Bug 924132 - Create a shared library that allows to deal with FB Data in a DataStore
