+++ apps/communications/contacts/fb_link.html
@@ -21,7 +21,9 @@
     <script  type="text/javascript" src="js/utilities/binary_search.js"></script>
     <script  defer type="text/javascript" src="/shared/js/lazy_loader.js"></script>
     <script  defer type="text/javascript" src="oauth2/js/oauth20.js"></script>
+    <script  defer type="text/javascript" src="/shared/js/fb/fb_request.js"></script>
     <script  defer type="text/javascript" src="js/fb/fb_data.js"></script>
+    <script  defer type="text/javascript" src="/shared/js/fb/fb_reader_utils.js"></script>
     <script  defer type="text/javascript" src="js/fb/fb_contact_utils.js"></script>
     <script  defer type="text/javascript" src="js/fb/fb_contact.js"></script>
     <script  defer type="text/javascript" src="/facebook/js/curtain.js"></script>
