+++ apps/communications/contacts/import.html
@@ -39,6 +39,11 @@
     <script defer type="text/javascript" src="js/importer_ui.js"></script>
     <script defer type="text/javascript" src="js/import_init.js"></script>
 
+    <!-- To include these files on the standard app
+      <link href="/shared/js/fb/fb_request.js">
+      <link href="/shared/js/fb/fb_data_reader.js">
+    -->
+
     <!-- Building Blocks -->
     <link href="/shared/style_unstable/progress_activity.css" rel="stylesheet">
     <link href="/shared/style/status.css" rel="stylesheet">
