+++ apps/communications/contacts/index.html
@@ -58,6 +58,8 @@
     <link href="/shared/js/contact2vcard.js">
     <link href="/shared/js/device_storage/get_storage_if_available.js">
     <link href="/shared/js/device_storage/get_unused_filename.js">
+    <link href="/shared/js/fb/fb_request.js">
+    <link href="/shared/js/fb/fb_data_reader.js">
     -->
 
     <link rel="import" href="elements/confirm.html">
