+++ apps/communications/contacts/js/import_init.js
@@ -39,9 +39,11 @@
     },
     'facebook': function(cb) {
       var files = [
-                   '/contacts/js/fb/fb_utils.js',
-                   '/contacts/js/fb/fb_contact_utils.js',
+                   '/shared/js/fb/fb_request.js',
                    '/contacts/js/fb/fb_data.js',
+                   '/contacts/js/fb/fb_utils.js',
+                   '/shared/js/fb/fb_reader_utils.js',
+                   '/contacts/js/fb/fb_contact_utils.js',
                    '/contacts/js/fb/fb_query.js',
                    '/contacts/js/fb/fb_contact.js',
                    '/facebook/js/facebook_connector.js',
