+++ apps/communications/dialer/index.html
@@ -48,8 +48,9 @@
     <script defer src="/shared/js/simple_phone_matcher.js"></script>
     <script defer src="/shared/js/settings_listener.js"></script>
     <script defer src="/contacts/js/utilities/confirm.js"></script>
-    <script defer src="/contacts/js/fb/fb_data.js"></script>
-    <script defer src="/contacts/js/fb/fb_contact_utils.js"></script>
+    <script defer src="/shared/js/fb/fb_request.js"></script>
+    <script defer src="/shared/js/fb/fb_data_reader.js"></script>
+    <script defer src="/shared/js/fb/fb_reader_utils.js"></script>
     <script defer src="/dialer/js/contacts.js"></script>
     <script defer src="/dialer/js/contact_data_manager.js"></script>
     <script defer src="/dialer/js/suggestion_bar.js"></script>
