+++ apps/communications/dialer/js/contacts.js
@@ -18,7 +18,7 @@ var _FbDataSearcher = function(variants) {
       var contact = result;
 
       if (contact) {
-        fb.utils.getMozContactByUid(contact.uid, function merge(e) {
+        fb.getMozContactByUid(contact.uid, function merge(e) {
           var devContact = e.target.result[0];
           var finalContact = fb.mergeContact(devContact, contact);
           successCb(finalContact, {
@@ -69,6 +69,12 @@ var _FbDataSearcher = function(variants) {
 
 var Contacts = {
 
+  _FB_FILES: [
+    '/shared/js/fb/fb_request.js',
+    '/shared/js/fb/fb_data_reader.js',
+    '/shared/js/fb/fb_reader_utils.js'
+  ],
+
   // The mozContact API stores a revision of its database that allow us to know
   // if we have a proper and updated contact cache.
   getRevision: function getRevision(callback) {
@@ -89,9 +95,8 @@ var Contacts = {
   },
 
   findByNumber: function findByNumber(number, callback) {
-    LazyLoader.load(['/contacts/js/fb/fb_data.js',
-                     '/contacts/js/fb/fb_contact_utils.js'],
-                  this._findByNumber.bind(this, number, callback));
+    LazyLoader.load(this._FB_FILES,
+                    this._findByNumber.bind(this, number, callback));
   },
 
   _findByNumber: function _findByNumber(number, callback) {
@@ -195,8 +200,7 @@ var Contacts = {
       callback(null);
     }
 
-    LazyLoader.load(['/contacts/js/fb/fb_data.js',
-                     '/contacts/js/fb/fb_contact_utils.js'], function() {
+    LazyLoader.load(this._FB_FILES, function() {
       for (var i = 0, length = contacts.length; i < length; i++) {
         if (fb.isFbLinked(contacts[i])) {
           var fbReq = fb.contacts.get(fb.getFriendUid(contacts[i]));
