+++ apps/communications/dialer/js/dialer.js
@@ -534,8 +534,9 @@ var dialerStartup = function startup(evt) {
 
     CallHandler.init();
     LazyL10n.get(function loadLazyFilesSet() {
-      LazyLoader.load(['/contacts/js/fb/fb_data.js',
-                       '/contacts/js/fb/fb_contact_utils.js',
+      LazyLoader.load(['/shared/js/fb/fb_request.js',
+                       '/shared/js/fb/fb_data_reader.js',
+                       '/shared/js/fb/fb_reader_utils.js',
                        '/shared/style/confirm.css',
                        '/contacts/js/utilities/confirm.js',
                        '/dialer/js/newsletter_manager.js',
