+++ apps/communications/dialer/oncall.html
@@ -38,8 +38,9 @@
     <script defer type="application/javascript" src="/dialer/js/handled_call.js"></script>
 
     <!-- Lazy load these:
-    <script defer type="application/javascript" src="/contacts/js/fb/fb_data.js"></script>
-    <script defer type="application/javascript" src="/contacts/js/fb/fb_contact_utils.js"></script>
+    <script defer type="application/javascript" src="/shared/js/fb/fb_request.js"></script>
+    <script defer type="application/javascript" src="/shared/js/fb/fb_data_reader.js"></script>
+    <script defer type="application/javascript" src="/shared/js/fb/fb_reader_utils.js"></script>
     -->
 
     <!-- incoming call / swiper on locked screen-->
