+++ apps/communications/facebook/fb_sync.html
@@ -6,14 +6,17 @@
     <script type="text/javascript" src="/shared/js/l10n.js"></script>
 
     <script src="/shared/js/async_storage.js"></script>
+    <script src="/shared/js/lazy_loader.js"></script>
     <script src="/shared/js/notification_helper.js"></script>
     <script src="/contacts/oauth2/js/parameters.js"></script>
     <script src="/contacts/js/utilities/config.js"></script>
     <script src="/contacts/js/utilities/image_square.js"></script>
+    <script src="/shared/js/fb/fb_request.js"></script>
     <script src="/contacts/js/fb/fb_data.js"></script>
     <script src="/contacts/js/fb/fb_contact.js"></script>
     <script src="/contacts/js/fb/fb_utils.js"></script>
     <script src="/contacts/js/import_utils.js"></script>
+    <script src="/shared/js/fb/fb_reader_utils.js"></script>
     <script src="/contacts/js/fb/fb_contact_utils.js"></script>
     <script src="js/fb_sync.js"></script>
     <script src="/contacts/js/fb/fb_init.js"></script>
