+++ apps/communications/facebook/js/sync_worker.js
@@ -1,7 +1,10 @@
 'use strict';
 
-importScripts('/contacts/js/fb/fb_query.js',
-              '/contacts/js/fb/fb_contact_utils.js', 'console.js');
+importScripts('/shared/js/fb/fb_request.js',
+              '/contacts/js/fb/fb_query.js',
+              '/contacts/js/fb/fb_contact_utils.js',
+              '/shared/js/fb/fb_reader_utils.js',
+              'console.js');
 
 (function(wutils) {
