+++ apps/communications/facebook/test/unit/facebook_connector_test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 requireApp('communications/contacts/js/fb/fb_contact_utils.js');
+require('/shared/js/fb/fb_reader_utils.js');
 requireApp('communications/contacts/js/fb/fb_contact.js');
 requireApp('communications/facebook/js/facebook_connector.js');
 requireApp('communications/facebook/test/unit/mock_fb_graph_data.js');
