+++ apps/communications/ftu/index.html
@@ -54,9 +54,11 @@
   <script defer type="text/javascript" src="/ftu/js/services_launcher.js"></script>
   <script defer type="text/javascript" src="/contacts/js/utilities/config.js"></script>
   <script defer type="text/javascript" src="/contacts/js/utilities/vcard_parser.js"></script>
+  <script defer type="text/javascript" src="/shared/js/fb/fb_request.js"></script>
   <script defer type="text/javascript" src="/contacts/js/fb/fb_data.js"></script>
   <script defer type="text/javascript" src="/contacts/js/fb/fb_utils.js"></script>
   <script defer type="text/javascript" src="/contacts/js/fb/fb_query.js"></script>
+  <script defer type="text/javascript" src="/shared/js/fb/fb_reader_utils.js"></script>
   <script defer type="text/javascript" src="/contacts/js/fb/fb_contact_utils.js"></script>
   <script defer type="text/javascript" src="/shared/js/async_storage.js"></script>
   <script defer type="text/javascript" src="/shared/js/icc_helper.js"></script>
